requirements.txt
flask
google-genai
python-dotenv
flask-cors
